pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-personal-agent-sdk"
version = "0.1.0"
description = "AI Personal Agent SDK for automation with OpenAI, Zapier, and security features"
authors = [
    {name = "Arun Kumar Singh", email = "arun250492@gmail.com"},
]
dependencies = [
    "openai>=1.0.0",
    "requests>=2.28.0",
    "cryptography>=41.0.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "oauth2client>=4.1.3",
    "google-api-python-client>=2.95.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "twilio>=8.2.0",  # for SMS/alarms
    "plyer>=2.1.0",   # for desktop notifications
    "psutil>=5.9.0",  # for system monitoring
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-personal-agent-sdk"
Repository = "https://github.com/yourusername/ai-personal-agent-sdk"
Issues = "https://github.com/yourusername/ai-personal-agent-sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_personal_agent_sdk*"]

[tool.setuptools.package-data]
"ai_personal_agent_sdk" = ["ui/static/*", "ui/templates/*"]